requirements.txt
fastapi
uvicorn
selenium
pandas
fhirpy
python-dotenv
webdriver-manager